requirements.txt
# LLM & AI
openai>=1.30.0
sentence-transformers>=3.0.0

# File parsing
pdfplumber>=0.11.0
python-docx>=1.1.0

# Data validation
pydantic>=2.0.0

# Evaluation metrics
scipy>=1.12.0
numpy>=1.26.0

# Environment
python-dotenv>=1.0.0

# Optional: Jupyter for evaluation notebook
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.2.0
